JavaDoc and small code quality fixes.

+++ elki-core-math/src/main/java/elki/math/statistics/dependence/mcde/MWPTest.java
@@ -115,7 +115,7 @@ public class MWPTest implements MCDETest<MWPTest.MWPRanking> {
 
       if(k > j) {
         double adjusted = (adjust + k) / (double) t;
-        correction += t * (t * t - 1);
+        correction += t * (t * (long) t - 1);
         if(correction < 0) {
           throw new ArithmeticException("Long overflow: too many ties (>2^10)");
         }
